adding authentication for upload/download/delete

+++ app/controllers/uploads_controller.rb
@@ -3,10 +3,25 @@ class UploadsController < ApplicationController
   def new
     @upload = Upload.new
     @uploads = Upload.all.order(created_at: :desc)
-    @admin = 1 if session[:user] && session[:user]["admin"] == 1
+  end
+
+  def show
+    if session[:user] == nil
+      render nothing: true, head: :unauthorized 
+      return
+    end
+
+    @upload = Upload.find params[:id]
+    send_file Rails.root.join('public', 'uploads', @upload["filename"])
+    
   end
 
   def create
+    if session[:user] == nil
+      render nothing: true, head: :unauthorized 
+      return
+    end
+
     uploaded_io = params[:upload][:filename]
     File.open(Rails.root.join('public', 'uploads', uploaded_io.original_filename), 'wb') do |file|
       file.write(uploaded_io.read)
@@ -20,7 +35,21 @@ class UploadsController < ApplicationController
   end
 
   def destroy
+    if session[:user] == nil
+      render nothing: true, head: :unauthorized 
+      return
+    elsif session[:user]["admin"] != 1 
+      render nothing: true, head: :forbidden
+      return
+    end 
+
     @upload = Upload.find params[:id]
+    
+    if @upload.user["username"] != session[:user]["username"]
+      render nothing: true, head: :forbidden 
+      return
+    end
+    
     @upload.destroy
 
     File.delete Rails.root.join('public', 'uploads', @upload[:filename])
